Renamed tests/handler.rs -> tests/uplink.rs. Added uplink with response tests.
